src/test/kotlin/com/github/rising3/crypto/HmacTest.kt
/**
 * MIT License
 *
 * Copyright (c) 2022 Michio Nakagawa <michio.nakagawa@gmail.com>
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
package com.github.rising3.crypto

import org.junit.jupiter.api.Test
import javax.crypto.spec.SecretKeySpec
import kotlin.test.assertContentEquals
import kotlin.test.assertTrue

class HmacTest {
    private val text = "Hello, world!"
    private val hmac = Hmac()
    private val key = SecretKeySpec(
        CryptoUtils.hexToByteArray("64186a9df99d5c953247307987f9007b610bd519182dd19507f9ad9a6c30d1ab"),
        "HmacSHA256"
    )

    @Test
    fun testHmac() {
        val mac = CryptoUtils.hexToByteArray("c16118213d1e4b512d83e0eaae396e970260638583eb4ac46345e64ba16ac7e6")
        val actual = hmac.mac(text.toByteArray(), key)

        assertContentEquals(mac, actual)
        assertTrue(hmac.verify(actual, text.toByteArray(), key))
    }
}
